+++ src/main/java/com/buuz135/industrial/tile/block/AnimalIndependenceSelectorBlock.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class AnimalIndependenceSelectorBlock extends CustomAreaOrientedBlock<AnimalIndependenceSelectorTile> {
 
     public AnimalIndependenceSelectorBlock() {
-        super("animal_independence_selector", AnimalIndependenceSelectorTile.class, Material.ROCK, 20 * 20, 20, RangeType.FRONT, 3, 1, false);
+        super("animal_independence_selector", AnimalIndependenceSelectorTile.class, Material.ROCK, 20 * 20, 20, RangeType.FRONT, 5, 1, true);
     }
 
     public void createRecipe() {