+++ src/main/java/com/buuz135/industrial/tile/block/CropEnrichMaterialInjectorBlock.java
@@ -12,7 +12,7 @@ import net.ndrei.teslacorelib.items.MachineCaseItem;
 public class CropEnrichMaterialInjectorBlock extends CustomAreaOrientedBlock<CropEnrichMaterialInjectorTile> {
 
     public CropEnrichMaterialInjectorBlock() {
-        super("crop_enrich_material_injector", CropEnrichMaterialInjectorTile.class, Material.ROCK, 400, 40, RangeType.FRONT, 1, 0, true);
+        super("crop_enrich_material_injector", CropEnrichMaterialInjectorTile.class, Material.ROCK, 400, 40, RangeType.FRONT, 11, 0, true);
     }
 
     public void createRecipe() {