+++ src/main/java/com/buuz135/industrial/tile/block/HydratorBlock.java
@@ -23,7 +23,7 @@ import java.util.Random;
 public class HydratorBlock extends CustomAreaOrientedBlock<HydratorTile> {
 
     public HydratorBlock() {
-        super("hydrator", HydratorTile.class, Material.ROCK, 1000, 10, RangeType.UP, 2, 0, false);
+        super("hydrator", HydratorTile.class, Material.ROCK, 1000, 10, RangeType.UP, 1, 0, true);
     }
 
     @Override