+++ src/main/java/com/buuz135/industrial/tile/mob/MobDuplicatorTile.java
@@ -17,6 +17,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Vec3d;
 import net.minecraftforge.fluids.IFluidTank;
 import net.minecraftforge.items.ItemStackHandler;
 import net.ndrei.teslacorelib.TeslaCoreLib;
@@ -98,12 +99,14 @@ public class MobDuplicatorTile extends WorkingAreaElectricMachine {
             if (experienceTank.getFluid() != null && experienceTank.getFluid().amount > essenceNeeded) {
                 entity = (EntityLiving) ((MobImprisonmentToolItem) stack.getItem()).getEntityFromStack(stack, this.world, BlockRegistry.mobDuplicatorBlock.enableExactCopy && exactCopy);
                 int tries = 20;
-                BlockPos random = blocks.get(this.world.rand.nextInt(blocks.size())).add(0.5, 0, 0.5);
+                Vec3d random = new Vec3d(blocks.get(this.world.rand.nextInt(blocks.size())));
+                random = random.addVector(0.5, 0, 0.5);
+                entity.setLocationAndAngles(random.x, random.y, random.z, world.rand.nextFloat() * 360F, 0);
                 entity.setUniqueId(UUID.randomUUID());
-                entity.setLocationAndAngles(random.getX(), random.getY(), random.getZ(), world.rand.nextFloat() * 360F, 0);
                 while (tries > 0 && !canEntitySpawn(entity)) {
-                    random = blocks.get(this.world.rand.nextInt(blocks.size()));
-                    entity.setLocationAndAngles(random.getX(), random.getY(), random.getZ(), world.rand.nextFloat() * 360F, 0);
+                    random = new Vec3d(blocks.get(this.world.rand.nextInt(blocks.size())));
+                    random = random.addVector(0.5, 0, 0.5);
+                    entity.setLocationAndAngles(random.x, random.y, random.z, world.rand.nextFloat() * 360F, 0);
                     --tries;
                 }