+++ src/main/java/com/buuz135/industrial/tile/world/EnergyFieldProviderTile.java
@@ -17,7 +17,7 @@ public class EnergyFieldProviderTile extends WorkingAreaElectricMachine {
 
     @Override
     public AxisAlignedBB getWorkingArea() {//(getRadius()*getRadius())/3 -1, (getRadius()*getRadius())/3 -1, (getRadius()*getRadius())/3 -1
-        return new AxisAlignedBB(this.pos.getX(), this.pos.getY(), this.pos.getZ(), this.pos.getX() + 1, this.pos.getY() + 1, this.pos.getZ() + 1).grow(getRadius() * 2, getRadius() * 2, getRadius() * 2);
+        return new AxisAlignedBB(this.pos.getX(), this.pos.getY(), this.pos.getZ(), this.pos.getX() + 1, this.pos.getY() + 1, this.pos.getZ() + 1).grow(getRadius() * 2 + 3, getRadius() * 2 + 3, getRadius() * 2 + 3);
     }
 
     public float consumeWorkEnergy(long consume) {